Migration of MapView to Jetpack Compose
This commits is the first part of the migration of the MapView Fragment to Jetpack Compose.

The commits also introduce the MVI architecture which will be used from now on to handle ui states and events. This refactoring is being done because MVI is a more natural fit for Jetpack Compose as it aligns well with it's declarative UI approach. it ensures a separation of ui state and operation.

This commit also migrates the bottom sheet modal to compose. changes will be made to ensure that it also uses the MVI architecture