Final part of MapView migration to Jetpack Compose.
A MVI architecture is used to simplify the use of the mapview using states and events.

This commit finishes the migration of the application to Jetpack Compose.

The last fragments are removed and MapView is now fully in Compose. Schedule data loading has been moved to the mapviewmodel for simplification

Navigation is now achieved using Compose Navigation and destinations.

Other changes include :
- Deletion of layout files as they are no longer required in jetpack compose
- Fix of the application icon
- Addition of a bus stop icon for the mapview